refactor(keysign): decompose loadTransaction into per-type builders & extract use cases (#4866)
Continues the JoinKeysignViewModel slim-down (tasks 2-4 of the issue plan):

- Decompose the ~745-line `loadTransaction` into per-transaction-type builders,
  each in its own file and independently testable:
  - `JoinSwapUiModelBuilder` (owns `buildSwapUiModel` + per-provider sub-branches)
  - `JoinDepositUiModelBuilder`
  - `JoinSendUiModelBuilder`
  `loadTransaction` is now a thin dispatch on payload type that applies the
  builder's `JoinKeysignVerifyResult`; the send branch's background hero/scan
  enrichment stays in the ViewModel (it launches into viewModelScope).
- Extract the suspend fee-resolution / EVM function-info helpers
  (`resolveJoinKeysignNetworkFee`, `resolveFallbackFeeAmount`,
  `getTransactionFunctionInfo`) into `JoinKeysignFeeResolver`.
- Extract the qBTC claim co-sign suspend work into `QbtcClaimCosignUseCase`
  (job launching, state transitions, and error mapping stay in the ViewModel).
- Fix `currentState` encapsulation: private `_currentState` MutableStateFlow +
  public read-only `StateFlow`, mirroring the existing `_dappMetadata` pattern.

Pure structural refactor — behavior is unchanged. The ViewModel drops from
2271 to 1094 lines and from 32 to 19 injected dependencies. `assembleDebug`,
`ktfmtFormat`, and the keysign unit tests pass.
